Teachable-Machine-Streamlit-main/requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tf2onnx>=1.15.0
opencv-python>=4.8.0
pillow>=10.0.0
plotly>=5.15.0
pydantic>=2.0.0
pyyaml>=6.0
albumentations>=1.3.0
grad-cam>=1.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-cov>=4.1.0
